Pull for initial 3.0.16 release.

General:

* Fix small bug in PsychPortAudio('Stop'): Did not block until true sound offset.
* Add new MouseMotionRecordingDemo.m: Shows how to record mouse movements et al. on Linux and Windows via use of enhanced keyboard queues.
* Various documentation updates, e.g., for Linux gamemode support, RetinaDisplays, GStreamer.
* Minor improvements and various cleanups.
* Datapixx updated to latest state from VPixx upstream.

Linux:

* Enable GstDeviceMonitors for nicer video capture device enumeration.
* Now requires at least GStreamer 1.4.0.
* Ubuntu 14.04-LTS or other distros older than Ubuntu 16.04-LTS no longer supported.
* Works with Octave 3.8 - 4.2. Octave 4.4 and 5.1 will be supported by NeuroDebian packages octave-psychtoolbox-3 or the Linux distributions octave-psychtoolbox-3 packages themselves.
* Now is developed and tested against Matlab R2019a and R2014b, no longer R2012a, but is expected - but not tested - to continue to work with R2012a.

Windows:

* Now absolutely requires the MSVC built GStreamer 1.16.0 for both Octave and Matlab. Screen() will no longer work without GStreamer 1.16.0 MSVC installed!
* May require Microsoft Visual Studio Runtime 2015-2019 redistributable to be installed.
* Keyboard queues can now record mouse/touchpad etc. motion as well, also raw motion, similar to Linux.
* Update PsychPortAudio's internal Portaudio library to latest upstream release with improvements and bug fixes to Windows WASAPI sound backend support.
* Now runs on and requires 64-Bit Octave 5.1, no longer on Octave 4.4.1.
* Now is tested and developed against Matlab R2019a, no longer R2012a, but should continue to work on older Matlab versions than R2019a.
* Crashes of Matlab in Java / GUI mode when using GStreamer functions like movie playback/writing/recording and video capture may be finally resolved! At least testing on Windows 10 with R2019a and R2014a did not experience any crashes anymore. Fingers crossed.
* Improved compatibility and easier setup of GStreamer 1.16.0 + Octave 5.1.0.

OSX:

* Improve robustness of Screen('ConfigureDisplay','Brightness') on Apples recent trainwrecks with NightShift support and such. Also add theoretical ability to control 'AutoBrightness' mode - in practice it doesn't seem to work at all.
* Now runs on and requires 64-Bit Octave 5.1, no longer on Octave 4.4.1.
* Now is tested and developed against Matlab R2019a, no longer R2012a, but should continue to work on older Matlab versions, possibly back to R2014b.

+++ PsychSourceGL/Source/Common/Screen/SCREENGetWindowInfo.c
@@ -514,7 +514,8 @@ PsychError SCREENGetWindowInfo(void)
         // Misc. window parameters:
         PsychSetStructArrayDoubleElement("StereoMode", 0, windowRecord->stereomode, s);
         PsychSetStructArrayDoubleElement("ImagingMode", 0, windowRecord->imagingMode, s);
-        PsychSetStructArrayDoubleElement("SpecialFlags", 0, windowRecord->specialflags, s);
+        // FIXME: Caution: specialflags is 64 bits, but double can only return about 53 low-bits correctly. 
+        PsychSetStructArrayDoubleElement("SpecialFlags", 0, (double) windowRecord->specialflags, s);
         PsychSetStructArrayDoubleElement("IsFullscreen", 0, (windowRecord->specialflags & kPsychIsFullscreenWindow) ? 1 : 0, s);
         PsychSetStructArrayDoubleElement("MultiSampling", 0, windowRecord->multiSample, s);
         PsychSetStructArrayDoubleElement("MissedDeadlines", 0, windowRecord->nr_missed_deadlines, s);
